refactor(agent): decompose AttackerAgent into Chunk collaborators
AttackerAgent had grown to 552 lines doing six jobs (orchestration,
per-chunk prompt assembly, caching, coverage, the sequential loop, and
the concurrent wavefront). Split per the single-responsibility rule into
cohesive Application\Agent\Chunk collaborators:

- ChunkContext (VO) + ChunkContextFactory — prompt assembly, slicing,
  cache-key/cacheability derivation.
- AttackerChunkCache — context-aware cache adapter + cache-hit hydration.
- ChunkCoverageRecorder — per-file coverage recording.
- SequentialChunkAnalyzer / ConcurrentChunkAnalyzer — the two analysis
  strategies.

AttackerAgent (now ~180 lines, mostly the constructor) builds these from
its existing injected dependencies, so the public constructor signature
and the DI wiring are unchanged and every analyze()-level test exercises
the collaborators unchanged. Behaviour is identical — full suite green.
