+++ src/Audit/Application/Agent/Chunk/ChunkContextFactory.php
@@ -1,0 +1,108 @@
+<?php
+
+/*
+ * This file is part of the vinceamstoutz/symfony-security-auditor package.
+ *
+ * (c) Vincent Amstoutz <vincent.amstoutz.dev@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace VinceAmstoutz\SymfonySecurityAuditor\Audit\Application\Agent\Chunk;
+
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Application\Agent\AttackerAnalysisRequest;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Application\Agent\AttackerContextPromptRenderer;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Application\Agent\RiskMarkerIndex;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Model\ProjectFile;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Port\AttackerPromptBuilderInterface;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Port\CodeSlicerInterface;
+
+/**
+ * Assembles the per-chunk system/user prompts (risk markers + cross-iteration
+ * preambles), slices each file to its security-relevant lines, and derives the
+ * cache key and cacheability. Shared by the sequential and concurrent chunk
+ * analyzers.
+ *
+ * @internal not part of the BC promise — see docs/versioning.md
+ */
+final readonly class ChunkContextFactory
+{
+    public function __construct(
+        private AttackerPromptBuilderInterface $attackerPromptBuilder,
+        private CodeSlicerInterface $codeSlicer,
+        private AttackerContextPromptRenderer $attackerContextPromptRenderer,
+    ) {}
+
+    /**
+     * @param list<ProjectFile> $chunk
+     */
+    public function create(array $chunk, AttackerAnalysisRequest $attackerAnalysisRequest, RiskMarkerIndex $riskMarkerIndex, bool $cacheIsContextAware): ChunkContext
+    {
+        $chunkMarkers = $riskMarkerIndex->forChunk($chunk);
+        $hasMarkers = [] !== $chunkMarkers;
+
+        $rejectedPreamble = [] !== $attackerAnalysisRequest->rejectedFindings
+            ? $this->attackerContextPromptRenderer->renderRejectedFindings($attackerAnalysisRequest->rejectedFindings)
+            : '';
+        $previousPreamble = [] !== $attackerAnalysisRequest->previousFindings
+            ? $this->attackerContextPromptRenderer->renderPreviousFindings($attackerAnalysisRequest->previousFindings)
+            : '';
+        $contextKey = '' === $rejectedPreamble && '' === $previousPreamble
+            ? ''
+            : hash('sha256', $rejectedPreamble."\0".$previousPreamble);
+        $cacheable = !$attackerAnalysisRequest->bypassCache && ('' === $contextKey || $cacheIsContextAware);
+
+        $slicedChunk = $this->sliceChunk($chunk);
+        $systemPrompt = $this->attackerPromptBuilder->buildSystemPrompt($slicedChunk);
+        $userMessage = $this->attackerPromptBuilder->buildUserMessage($slicedChunk, $attackerAnalysisRequest->symfonyMapping);
+
+        if ($hasMarkers) {
+            $userMessage = $this->attackerContextPromptRenderer->renderRiskMarkers($chunkMarkers)."\n\n".$userMessage;
+        }
+
+        if ('' !== $rejectedPreamble) {
+            $userMessage = $rejectedPreamble."\n\n".$userMessage;
+        }
+
+        if ('' !== $previousPreamble) {
+            $userMessage = $previousPreamble."\n\n".$userMessage;
+        }
+
+        return new ChunkContext($systemPrompt, $userMessage, $contextKey, $cacheable);
+    }
+
+    /**
+     * Replaces each file in the chunk with a version whose content is sliced
+     * down to security-relevant lines. The slicer preserves the original line
+     * count by replacing elided lines with a `// elided` placeholder, so the
+     * line-numbering protocol in the prompt stays accurate against the source.
+     *
+     * @param list<ProjectFile> $chunk
+     *
+     * @return list<ProjectFile>
+     */
+    private function sliceChunk(array $chunk): array
+    {
+        $sliced = [];
+        foreach ($chunk as $file) {
+            $newContent = $this->codeSlicer->slice($file);
+
+            if ($newContent === $file->content()) {
+                $sliced[] = $file;
+
+                continue;
+            }
+
+            $sliced[] = ProjectFile::create(
+                relativePath: $file->relativePath(),
+                absolutePath: $file->absolutePath(),
+                content: $newContent,
+            );
+        }
+
+        return $sliced;
+    }
+}
